+++ app/controllers/instructions_controller.rb
@@ -1,5 +1,6 @@
 class InstructionsController < ApplicationController
   def new
+    session_notice(:danger, 'You have to be logged in!', login_path) unless logged_in?
     @recipe = Recipe.find(params[:recipe_id])
     @instruction = @recipe.instructions.build
   end
@@ -7,6 +8,7 @@ class InstructionsController < ApplicationController
   def create
     @recipe = Recipe.find(params[:recipe_id])
     @instruction = @recipe.instructions.build(instruction_params)
+    @instruction.user = current_user
 
     if @instruction.save
       redirect_to @recipe
@@ -16,7 +18,11 @@ class InstructionsController < ApplicationController
   end
 
   def edit
-    @instruction = instruction.find(params[:id])
+    session_notice(:danger, 'You have to be logged in!', login_path) unless logged_in?
+    @instruction = Instruction.find(params[:id])
+    if logged_in?
+      session_notice(:danger, 'Wrong User') unless equal_with_current_user?(@instruction.user)
+    end
     @recipe = @instruction.recipe
   end
 
@@ -32,9 +38,14 @@ class InstructionsController < ApplicationController
   end
 
   def destroy
+    session_notice(:danger, 'You must be logged in!', login_path) unless logged_in?
     instruction = Instruction.find(params[:id])
-    instruction.destroy
-    redirect_to instruction.recipe
+    if equal_with_current_user?(instruction.user)
+      instruction.destroy
+      redirect_to instruction.recipe
+    else
+      session_notice(:danger, 'Wrong User')
+    end
   end
 
   private
